+++ test/custom_navigation_custom_features_test.dart
@@ -748,12 +748,14 @@ void main() {
 
     await tester.pumpWidget(
       MaterialApp(
-        theme: ThemeData(useMaterial3: true),
+        theme: ThemeData(
+          useMaterial3: true,
+          appBarTheme: AppBarTheme(shape: fillShape),
+        ),
         home: Scaffold(
           body: CustomNavigationRail(
             selectedIndex: 0,
             destinationFillRegion: NavigationDestinationRegion.full,
-            destinationFillShape: fillShape,
             destinations: const <NavigationRailDestination>[
               NavigationRailDestination(
                 icon: Icon(Icons.home_outlined),